feat: 쿠폰 관리 훅 추가 및 App.tsx에서 쿠폰 관련 로직 통합
- useCoupons 훅을 추가하여 쿠폰 상태 관리 및 적용 로직을 분리
- App.tsx에서 쿠폰 관련 상태 및 계산 로직을 useCoupons 훅으로 통합
- 기존 쿠폰 관련 로직 주석 처리 및 로컬 스토리지 업데이트 로직 수정
